+++ server/src/helpers/events.ts
@@ -1,0 +1,3 @@
+import { TextDocument } from "vscode-languageserver";
+export type DocumentEvent = (document: TextDocument) => any;
+export type ConfigEvent = (settings: any) => any;