+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Split into Client and Language Server.
+
+### Changed
+- Big structural change to run in client/server mode
+- Fixed `package.json` information
 
 ## [0.1.0] - 2019-04-23
 - Initial release